+++ src/com/asiainfo/ocmanager/rest/bean/service/instance/SparkServiceInstanceQuotaBean.java
@@ -24,7 +24,7 @@ public class SparkServiceInstanceQuotaBean extends ServiceInstanceQuotaBean {
 
 	public final static String YARNQUEUEQUOTA = "yarnQueueQuota";
 
-	private long yarnQueueQuota;
+	private double yarnQueueQuota;
 
 	public SparkServiceInstanceQuotaBean() {
 
@@ -157,11 +157,11 @@ public class SparkServiceInstanceQuotaBean extends ServiceInstanceQuotaBean {
 
 	}
 
-	public long getYarnQueueQuota() {
+	public double getYarnQueueQuota() {
 		return yarnQueueQuota;
 	}
 
-	public void setYarnQueueQuota(long yarnQueueQuota) {
+	public void setYarnQueueQuota(double yarnQueueQuota) {
 		this.yarnQueueQuota = yarnQueueQuota;
 	}
